refactor for better organization
