+++ index.ts
@@ -7,7 +7,7 @@ export type { TokenSet } from './src/types/token-set.ts';
 export type { PublicTokenSet } from './src/types/public-token-set.ts';
 export type { PublicClassSet } from './src/types/public-class-set.ts';
 export type { PublicApiTokens } from './src/types/public-api-tokens.js';
-export type { PublicApiGroup } from './src/types/public-api-group.ts';
+export type { PublicApiTokenGroup } from './src/types/public-api-token-group.js';
 export type { PublicApiToken } from './src/types/public-api-token.ts';
 export type { PublicApiClasses } from './src/types/public-api-classes.ts';
 export type { PublicApiClassGroup } from './src/types/public-api-class-group.ts';
