+++ src/plugins/build-style-dictionary-plugin.mts
@@ -2,402 +2,33 @@ import { readFile } from 'fs/promises';
 import { sync } from 'glob';
 import { Plugin } from 'vite';
 import path from 'path';
-import StyleDictionary, {
-  Config,
-  PlatformConfig,
-  Token,
-  TransformedTokens,
-} from 'style-dictionary';
+import StyleDictionary from 'style-dictionary';
 import { formattedVariables, sortByName } from 'style-dictionary/utils';
 import { getTransforms, register } from '@tokens-studio/sd-transforms';
 import { TokenConfig } from '../types/token-config.js';
-import { TokenSet } from '../types/token-set.js';
-import { Breakpoint } from '../types/breakpoint.js';
-import { ReferenceTokenSet } from '../types/reference-token-set.js';
-import {
-  generateAssetsCss,
-  fixAssetsUrlValue,
-} from './shared/assets-utils.mjs';
-import { PublicTokenSet } from '../types/public-token-set.js';
 import { PublicApiTokens } from '../types/public-api-tokens.js';
-import { PublicApiGroup } from '../types/public-api-group.js';
-import { PublicApiToken } from '../types/public-api-token.js';
 import { PublicApiClasses } from '../types/public-api-classes.js';
-import { PublicApiClassGroup } from '../types/public-api-class-group.js';
-import { PublicApiClass } from '../types/public-api-class.js';
-
-interface SkyStyleDictionaryConfig extends Config {
-  platforms: {
-    css: PlatformConfig;
-    json?: PlatformConfig;
-  };
-}
-
-interface GeneratedFile {
-  output: unknown;
-  destination: string | undefined;
-  breakpoint?: Breakpoint;
-}
-
-interface SkyTokenOptions {
-  assetsBasePath: string;
-  generateUrlAtProperties?: boolean;
-  showDescriptions?: boolean;
-  selectorPrefix: string;
-}
-
-const DEFAULT_SD_CONFIG: SkyStyleDictionaryConfig = {
-  preprocessors: ['tokens-studio'],
-  platforms: {
-    css: {
-      transformGroup: 'custom/tokens-studio',
-      options: {
-        outputReferences: true,
-        showFileHeader: false,
-      },
-      buildPath: `dist/`,
-    }
-  },
-};
-
-function isUrlToken(token: Token): boolean {
-  return (
-    token.$extensions?.['com.blackbaud.developer.sky-token-format'] === 'url'
-  );
-}
-
-function buildPublicApiGroups(
-  allTokens: Token[],
-  tokenTree: TransformedTokens,
-): PublicApiTokens {
-  const result: PublicApiTokens = {};
-
-  for (const token of allTokens) {
-    // Walk the token's path through the raw token tree, collecting any ancestor
-    // nodes that declare a groupName extension. This becomes an ordered list of
-    // groups from outermost to innermost (e.g. [Colors, Text Colors]).
-    const groupPath: { groupName: string; description?: string }[] = [];
-    let current: TransformedTokens = tokenTree;
-
-    for (const segment of token.path) {
-      current = current[segment];
-      const node = current as unknown as Token;
-      if (node.$extensions?.groupName) {
-        groupPath.push({
-          groupName: node.$extensions.groupName as string,
-          description: node.$description,
-        });
-      }
-    }
-
-    const tokenEntry: PublicApiToken = {
-      name: (token.$extensions?.readableName as string) ?? token.name,
-      cssProperty: `--${token.name}`,
-    };
-
-    if (token.$description) {
-      tokenEntry.description = token.$description;
-    }
-
-    if (token.$extensions?.deprecated) {
-      tokenEntry.deprecated = token.$extensions.deprecated as string;
-    }
-
-    // Tokens with no group ancestry go to the top-level tokens array.
-    if (groupPath.length === 0) {
-      result.tokens ??= [];
-      result.tokens.push(tokenEntry);
-    } else {
-      // Walk (or create) the nested group structure, then place the token in
-      // the innermost group.
-      result.groups ??= [];
-      let currentGroups = result.groups;
-
-      for (let i = 0; i < groupPath.length; i++) {
-        const { groupName, description } = groupPath[i];
-        let group = currentGroups.find((g) => g.groupName === groupName);
-        if (!group) {
-          group = { groupName };
-          currentGroups.push(group);
-        }
-        if (description && !group.description) {
-          group.description = description;
-        }
-        if (i < groupPath.length - 1) {
-          // Intermediate group: descend into its subgroups.
-          group.groups ??= [];
-          currentGroups = group.groups;
-        } else {
-          // Leaf group: append the token here.
-          group.tokens ??= [];
-          group.tokens.push(tokenEntry);
-        }
-      }
-    }
-  }
-
-  return result;
-}
-
-function mergePublicApiGroupArrays(
-  target: PublicApiGroup[],
-  source: PublicApiGroup[],
-): void {
-  for (const srcGroup of source) {
-    const existing = target.find((g) => g.groupName === srcGroup.groupName);
-    if (existing) {
-      if (srcGroup.description && !existing.description) {
-        existing.description = srcGroup.description;
-      }
-      if (srcGroup.tokens) {
-        existing.tokens ??= [];
-        for (const token of srcGroup.tokens) {
-          if (
-            !existing.tokens.some((t) => t.cssProperty === token.cssProperty)
-          ) {
-            existing.tokens.push(token);
-          }
-        }
-      }
-      if (srcGroup.groups) {
-        existing.groups ??= [];
-        mergePublicApiGroupArrays(existing.groups, srcGroup.groups);
-      }
-    } else {
-      target.push(srcGroup);
-    }
-  }
-}
-
-function mergePublicApiResults(
-  target: PublicApiTokens,
-  source: PublicApiTokens,
-): void {
-  if (source.tokens) {
-    target.tokens ??= [];
-    for (const token of source.tokens) {
-      if (!target.tokens.some((t) => t.cssProperty === token.cssProperty)) {
-        target.tokens.push(token);
-      }
-    }
-  }
-  if (source.groups) {
-    target.groups ??= [];
-    mergePublicApiGroupArrays(target.groups, source.groups);
-  }
-}
-
-function mergePublicApiClassGroupArrays(
-  target: PublicApiClassGroup[],
-  source: PublicApiClassGroup[],
-): void {
-  for (const srcGroup of source) {
-    const existing = target.find((g) => g.groupName === srcGroup.groupName);
-    if (existing) {
-      if (srcGroup.description && !existing.description) {
-        existing.description = srcGroup.description;
-      }
-      if (srcGroup.classes) {
-        existing.classes ??= [];
-        for (const cls of srcGroup.classes) {
-          if (!existing.classes.some((c: PublicApiClass) => c.cssClass === cls.cssClass)) {
-            existing.classes.push(cls);
-          }
-        }
-      }
-      if (srcGroup.groups) {
-        existing.groups ??= [];
-        mergePublicApiClassGroupArrays(existing.groups, srcGroup.groups);
-      }
-    } else {
-      target.push(srcGroup);
-    }
-  }
-}
-
-function mergePublicApiClassesResults(
-  target: PublicApiClasses,
-  source: PublicApiClasses,
-): void {
-  if (source.classes) {
-    target.classes ??= [];
-    for (const cls of source.classes) {
-      if (!target.classes.some((c) => c.cssClass === cls.cssClass)) {
-        target.classes.push(cls);
-      }
-    }
-  }
-  if (source.groups) {
-    target.groups ??= [];
-    mergePublicApiClassGroupArrays(target.groups, source.groups);
-  }
-}
-
-function getMediaQueryMinWidth(breakpoint: Breakpoint): string {
-  switch (breakpoint) {
-    case 'xs':
-    default:
-      return '0px';
-    case 's':
-      return '768px';
-    case 'm':
-      return '992px';
-    case 'l':
-      return '1200px';
-  }
-}
-
-function generatePublicClassGroupCss(
-  group: PublicApiClassGroup,
-  indent: string,
-): string {
-  const lines: string[] = [];
-
-  lines.push(`${indent}/* ${group.groupName} */`);
-  if (group.description) {
-    lines.push(`${indent}/* ${group.description} */`);
-  }
-  lines.push('');
-
-  if (group.classes) {
-    for (const cls of group.classes) {
-      if (cls.description) {
-        lines.push(`${indent}/* ${cls.description} */`);
-      }
-      lines.push(`${indent}.${cls.cssClass} {`);
-      if (cls.cssProperties) {
-        for (const [prop, value] of Object.entries(cls.cssProperties)) {
-          lines.push(`${indent}  ${prop}: ${value};`);
-        }
-      }
-      lines.push(`${indent}}`);
-      lines.push('');
-    }
-  }
-
-  if (group.groups) {
-    for (const subgroup of group.groups) {
-      lines.push(generatePublicClassGroupCss(subgroup, indent));
-    }
-  }
-
-  return lines.join('\n');
-}
-
-function generatePublicClassesCss(
-  publicApiClasses: PublicApiClasses,
-  selector: string,
-): string {
-  const lines: string[] = [];
-  const indent = '  ';
-
-  lines.push(`${selector} {`);
-
-  if (publicApiClasses.classes) {
-    for (const cls of publicApiClasses.classes) {
-      if (cls.description) {
-        lines.push(`${indent}/* ${cls.description} */`);
-      }
-      lines.push(`${indent}.${cls.cssClass} {`);
-      if (cls.cssProperties) {
-        for (const [prop, value] of Object.entries(cls.cssProperties)) {
-          lines.push(`${indent}  ${prop}: ${value};`);
-        }
-      }
-      lines.push(`${indent}}`);
-      lines.push('');
-    }
-  }
-
-  if (publicApiClasses.groups) {
-    for (const group of publicApiClasses.groups) {
-      lines.push(generatePublicClassGroupCss(group, indent));
-    }
-  }
-
-  lines.push('}');
-  lines.push('');
-
-  return lines.join('\n');
-}
-
-function collectGroupCssProperties(
-  group: PublicApiGroup,
-  result: Set<string>,
-): void {
-  if (group.tokens) {
-    for (const token of group.tokens) {
-      result.add(token.cssProperty);
-    }
-  }
-  if (group.groups) {
-    for (const subgroup of group.groups) {
-      collectGroupCssProperties(subgroup, result);
-    }
-  }
-}
-
-function collectPublicTokenCssProperties(
-  api: PublicApiTokens,
-  result = new Set<string>(),
-): Set<string> {
-  if (api.tokens) {
-    for (const token of api.tokens) {
-      result.add(token.cssProperty);
-    }
-  }
-  if (api.groups) {
-    for (const group of api.groups) {
-      collectGroupCssProperties(group, result);
-    }
-  }
-  return result;
-}
-
-function extractVarReferences(value: string): string[] {
-  return [...value.matchAll(/var\((--[^,)]+)/g)].map((m) => m[1].trim());
-}
-
-function validatePublicClassesCssProperties(
-  publicApiClasses: PublicApiClasses,
-  knownCssProperties: Set<string>,
-  setName: string,
-): void {
-  const errors: string[] = [];
-
-  function checkClass(cssClass: string, cssProperties: Record<string, string> | undefined): void {
-    if (!cssProperties) return;
-    for (const value of Object.values(cssProperties)) {
-      for (const ref of extractVarReferences(value)) {
-        if (!knownCssProperties.has(ref)) {
-          errors.push(`  .${cssClass}: "${ref}" is not defined in publicTokens`);
-        }
-      }
-    }
-  }
-
-  function walkGroup(group: PublicApiClassGroup): void {
-    for (const cls of group.classes ?? []) {
-      checkClass(cls.cssClass, cls.cssProperties);
-    }
-    for (const subgroup of group.groups ?? []) {
-      walkGroup(subgroup);
-    }
-  }
-
-  for (const cls of publicApiClasses.classes ?? []) {
-    checkClass(cls.cssClass, cls.cssProperties);
-  }
-  for (const group of publicApiClasses.groups ?? []) {
-    walkGroup(group);
-  }
-
-  if (errors.length) {
-    throw new Error(
-      `Invalid CSS custom property references in "${setName}":\n${errors.join('\n')}`,
-    );
-  }
-}
+import { fixAssetsUrlValue } from './shared/assets-utils.mjs';
+import {
+  buildPublicApiGroups,
+  collectPublicTokenCssProperties,
+  mergePublicApiResults,
+} from './shared/public-api-tokens-utils.mjs';
+import {
+  generatePublicClassesCss,
+  mergePublicApiClassesResults,
+  validatePublicClassesCssProperties,
+} from './shared/public-api-classes-utils.mjs';
+import {
+  addAssetsCss,
+  GeneratedFile,
+  getBaseDictionaryConfig,
+  getMediaQueryMinWidth,
+  getPublicDictionaryConfig,
+  getReferenceDictionaryConfig,
+  isUrlToken,
+  SkyTokenOptions,
+} from './shared/style-dictionary-config.mjs';
 
 async function generateDictionaryFiles(
   tokenConfig: TokenConfig,
@@ -524,124 +155,6 @@ async function generateDictionaryFiles(
   });
 
   return { tokenFiles, publicApiFiles, publicApiJsonFiles, publicClassFiles, publicClassJsonFiles };
-}
-
-function getBaseDictionaryConfig(
-  rootPath: string,
-  tokenSet: TokenSet,
-  skyOptions: SkyTokenOptions,
-): SkyStyleDictionaryConfig {
-  const config = structuredClone(DEFAULT_SD_CONFIG);
-
-  config.source = [`${rootPath}${tokenSet.path}`];
-
-  const cssOptions = (config.platforms.css.options ??= {});
-
-  Object.assign(cssOptions, {
-    skyOptions,
-    selector: tokenSet.selector,
-  });
-
-  config.platforms.css.files = [
-    {
-      destination: `${tokenSet.name}/${tokenSet.name}.css`,
-      format: 'css/alphabetize-variables',
-      filter: (token) => token.isSource,
-    },
-  ];
-
-  return config;
-}
-
-function getReferenceDictionaryConfig(
-  rootPath: string,
-  tokenSet: TokenSet,
-  referenceTokenSet: ReferenceTokenSet,
-  skyOptions: SkyTokenOptions,
-): SkyStyleDictionaryConfig {
-  const config = structuredClone(DEFAULT_SD_CONFIG);
-  config.source = [`${rootPath}${tokenSet.path}`];
-  config.include = [`${rootPath}${referenceTokenSet.path}`];
-
-  const cssOptions = (config.platforms.css.options ??= {});
-
-  Object.assign(cssOptions, {
-    skyOptions,
-    selector: `${tokenSet.selector}${referenceTokenSet.selector || ''}`,
-  });
-
-  config.platforms.css.files = [
-    {
-      destination: `${tokenSet.name}/${referenceTokenSet.name}.css`,
-      format: 'css/alphabetize-variables',
-      filter: (token) => !token.isSource,
-    },
-  ];
-
-  return config;
-}
-
-function getPublicDictionaryConfig(
-  rootPath: string,
-  tokenSet: TokenSet,
-  publicTokenSet: PublicTokenSet,
-  skyOptions: SkyTokenOptions,
-): SkyStyleDictionaryConfig {
-  const config = structuredClone(DEFAULT_SD_CONFIG);
-  config.source = [`${rootPath}${tokenSet.path}`];
-  config.include = [
-    `${rootPath}${publicTokenSet.path}`,
-    ...tokenSet.referenceTokens.map(
-      (referenceTokenSet) => `${rootPath}${referenceTokenSet.path}`,
-    ),
-  ];
-
-  const cssOptions = (config.platforms.css.options ??= {});
-
-  Object.assign(cssOptions, {
-    skyOptions: { ...skyOptions, showDescriptions: true },
-    selector: tokenSet.selector,
-  });
-
-  config.platforms.css.files = [
-    {
-      destination: `${tokenSet.name}/${publicTokenSet.name}.css`,
-      format: 'css/alphabetize-variables',
-      filter: (token) =>
-        path.normalize(token.filePath) ===
-        path.normalize(`${rootPath}${publicTokenSet.path}`),
-    },
-  ];
-
-  config.platforms.json = {
-    transformGroup: 'custom/tokens-studio',
-    buildPath: 'dist/',
-    files: [
-      {
-        destination: `${tokenSet.name}/${publicTokenSet.name}.json`,
-        format: 'json/public-api',
-        filter: (token) =>
-          path.normalize(token.filePath) ===
-          path.normalize(`${rootPath}${publicTokenSet.path}`),
-      },
-    ],
-  };
-
-  return config;
-}
-
-async function addAssetsCss(
-  tokenConfig: TokenConfig,
-  basePath: string,
-  fileContents: string,
-): Promise<string> {
-  const assetsCss = await generateAssetsCss(basePath, tokenConfig.projectName);
-
-  if (assetsCss) {
-    fileContents = `${assetsCss}\n\n${fileContents}`;
-  }
-
-  return fileContents;
 }
 
 export function buildStyleDictionaryPlugin(tokenConfig: TokenConfig): Plugin {
@@ -847,7 +360,7 @@ ${variables}
       if (publicApiJsonData.groups || publicApiJsonData.tokens) {
         this.emitFile({
           type: 'asset',
-          fileName: 'bundles/public-api.json',
+          fileName: 'bundles/public-api-tokens.json',
           source: JSON.stringify(publicApiJsonData, null, 2),
         });
       }
