+++ src/types/generated-file.ts
@@ -1,0 +1,7 @@
+import { Breakpoint } from './breakpoint.js';
+
+export interface GeneratedFile {
+  output: unknown;
+  destination: string | undefined;
+  breakpoint?: Breakpoint;
+}
