+++ src/types/public-api-class-group.ts
@@ -5,4 +5,5 @@ export interface PublicApiClassGroup {
   description?: string;
   groups?: PublicApiClassGroup[];
   classes?: PublicApiClass[];
+  groupImageToken?: string;
 }
