+++ src/types/public-api-class.ts
@@ -1,7 +1,7 @@
 export interface PublicApiClass {
   name: string;
   cssClass: string;
+  cssProperties: Record<string, string>;
   description?: string;
   deprecated?: string;
-  cssProperties?: Record<string, string>;
 }
