+++ src/types/sky-style-dictionary-config.ts
@@ -1,0 +1,8 @@
+import { Config, PlatformConfig } from 'style-dictionary';
+
+export interface SkyStyleDictionaryConfig extends Config {
+  platforms: {
+    css: PlatformConfig;
+    json?: PlatformConfig;
+  };
+}
