+++ src/types/sky-token-options.ts
@@ -1,0 +1,6 @@
+export interface SkyTokenOptions {
+  assetsBasePath: string;
+  generateUrlAtProperties?: boolean;
+  showDescriptions?: boolean;
+  selectorPrefix: string;
+}
